Add external controller migration export
We are using the same mechanism as for imports, in this case we fill the
description model with the ExternalController as retrieved from the
dqlite database.
Also another method for retrieving the list of model uuids for one external
controller was added to the state, needed to fill the description model.