Add 'integrations' to juju status
Now that we're moving over to the term 'integration' instead of
'relation', the option '--integrations' should be added alongside
'--relations'

Also, in tabular output use the term 'integration' instead of
'relation'. Keep 'relation' in json/yaml output as this would be a
breaking change.

This should probably be changed to follow suit in juju 4