Move the output package to core/output
The following moves the cmd/output package to core/output. The output
package has no juju/juju dependencies other than core/status. So moving
this package holds it to a higher standard. The core package enforces that
the code in here shouldn't depend on juju/juju code outside of the core
packages.