Take sqlair preparing out of contexts
We should avoid doing as much processing inside transactions as
possible. Especially when it could potentially fail for reasons
independent of the database. This could cause many pointless repetitions
of the transaction.

This required restructuring the package to rely far less on utility
functions. As a consequence, there is also much less redunancy in the
package

Write some more tests, and ensure that a completed upgrade is present
before every test, to ensure completed upgrades don't interfer with new
ones.

Also, re-write AllProvisionedControllersReady to perform business logic
in the database