Remove GUIArgs and GetChanges bundle features
GetChanges has been deprecated in favour of GetChangesMapArgs, which
doesn't rely clients being aware of a magical ordering. Remove
GetChanges now that we have a major version bump

This means we can remove GUIArgs alongside which is now dead code.

As a consequence, we will need to ensure that our clients such as
pylibjuju only use GetChangesMapArgs. At the moment, pylibjuju supports
both