Fixes TestRemoveMachine in the firewaller worker test suite.
This fails intermittently, because the changes channels are buffered
instead of using proper synchronisation in test orchestrations.

There is not way to tell which event the worker's loop will pick up
first of the changes enqueued for unit, application and machine.

We could add AnyTimes() here, but there is no point deliberately
orchestrating a maybe scenario into the test.