Use type to map rows from queries to core types
This commit replaces the use of the internal function in state `controllerInfoFromRows`
which was used for mapping returned rows from queries to core types, now
the type ExternalControllers (which is a type alias to a slice of
ExternalController and contains db tags) has a method for directly being
mapped to the core type.

Also the ControllersForModels query is modified to include a sub-select
when selecting the list of models.