Increase rev of juju-qa-test in bundle
This fixes a failure in run_deploy_exported_charmhub_bundle_with_fixed_revisions

To be honest I don't entire understand why this resolves the issue.
Perhaps something has broken with the old revision?

The fixed revision of juju-qa-jenkins was really old, but the test still
functions as expected if we increase this rev, so long as we keep it
fixed