Move convinient factory methods for TxRunner and WatchableDB to core
We used to have factory methods for creating `TxnRunnerFactory`s on
domain/state.go but since we are now also going to be using them in the
migration package, it's better to have them at core.
NewTxnRunnerFactoryForNamespace will now be in the core/database package
(core/database/runner.go), and NewTxnRunnerFactory will be on
core/changestream package (core/changestream/eventsource.go) since it
depends on WatchableDB and we cannot have cyclic references on it from
core/database.