Backport the setting of foreign keys via pragma
The following forces the setting of the foreign keys pragma upon
opening the database. This ensures that we never attempt to run
a db without setting this key functionality.

This potentially can cause issues with existing foreign keys, but
only leases are avaiable in this release. The foreign key usage is
extremely small. It is used in two places:

 1. The lease type id, which ensures that the lease is either a
    singular-controller or an application-leader. Neither of those
    are allowed to be user defined, so we can ensure that it's
    either one of those.

 2. The second type is based on the pinned lease. This also requires
    no user input and is set by code. This is a direct link to the
    lease table uuid.

The change log tables are not used or set, so we can ignore that
schema. This should have been removed from the release as we ended
up moving that to the main branch.