Allow retrying to locate the model
To make the juju wait-for model more flexible, it allows the retrying
of the model if the model isn't located. We have to do this in two
stages. The first one is when we start, if the controller isn't
located we have to retry until it might be (maybe during a bootstrap)
and then we we have to try again if the controller is located, but
the model hasn't been created.