Reset the app status once we've run the query
The following resets the app info status to the prior value once
the query has run. This is because the order of the deltas is not
guaranteed. Prior to this change we would only ever run the query
if the application had changed, not if the units had changed.

In workflows that don't chain the wait-for commands together and
before the application was created then you would get the
following (see below) where the appInfo would be somewhat near the
end of the deltas messages we recieved:

 - unitInfo
 - unitInfo
 - machineInfo
 - machineInfo
 - appInfo

If you attempt to do this before the application is created, then
the ordering is different, as one would expect:

 - appInfo
 - unitInfo
 - machineInfo
 - unitInfo
 - machineInfo

The problem here, is that the status of the appInfo was overridden
from `unset` to what the derived status of the units status. The
fix for now is to put the value back after each run request. This
then allows us to monitor the unset along with the status if it is
set.

The question is, why now? Why did this break now? It's clear that
a race was occuring, between the wait-for and the application being
registered. Either something has got slower or something has got
faster. The problem was that the wait-for wasn't robust enough to
manage the out of order appInfo. We now don't care if it is.