Introduction of a service registry
The following introduces a service registry for the api server. The
new registry allows a apiserver to peal off a service based on the
criteria that they require. The upside of this, is that we now have
the ability to implement various testing strategies. Either we can
use the controller suite db package, or we can use mocking depending
on the use cases.

We can also hide the DeleteDB from the facade context and work to
remove the controller db from the facade context is next on the list.
We can do that once the importing logic has no dependencies on the
state package.

At no point will it be required to use the db from the api server,
limiting the surface area for each facade.