Remove all external controllers references from legacy state
This patch removes everything that references external controllers from
the legacy mongodb state, thus finishing the move from mongodb to dqlite
for this specific feature.

Note: the remote applications cod had a small business logic in the
state to clean-up the external controller when the removed app was the
last one on that (external) controller. This has been removed and will
be reimplemented later using a new subsystem that will deal with orphan
resources like this one.