Remove controllerport worker
The following removes the controllerport worker. This is no longer
required as we're making the controller-config controller-api-port
immutable. This means we no longer need a worker to track the
changes.

Obviously, there will be documentation changes, but we've now removed
some inter-dependencies between the controller port worker and other
workers that required to watch if it was bouncing or not.