Late bind the services registry
Only create the services registry when required. This means that
we don't need to create lots of registries only to throw them away.
This is only per connection, so if the connection doesn't need an
API that requires the services, then it won't be used. Speeding
up any connections with no db access.