Pass in the controller config into set api host ports
The following passes in the controller config to set the api host
ports. Unfortunately, this is a requirement, and we can't wrap it
as we need the logic to be in state. The following is a partial fix
to allow controller config to be passed.

There are a lot more call sites to be fixed here.