Add watcher to upgrade that notifies when upgrade can start
Instead of, as we are with mongo, watching for any changes and then
requiring the consumer to verify it's an actionable change, create a new
watcher extending NamespaceWatcher which does this verification for you

This combines Watch and AllProvisionedControllersReady from
state/upgrade

Test this by mocking the underlying watcher. Ideally, we would have some
tests watching a real database, but we are not able to do this at the
moment.