Service factory worker
The following moves the service factory into a direct worker. The
worker can get all the correct dbs and then forward them to the output
of the manifold. The apiserver or other workers can then peel off the
services as required.

This is fundamental change between the state and statepool infrastructure
we had previously. There shouldn't be a requirement for an API facade to
use a db directly. Currently there is one exception to this and that is
import and export. That can easily be moved into the services once the
legacy state package has been removed.

As the service factory is just a light weight wrapper around all the
various domain services the cost should be relatively small. Having
said that, there are opportunities to pool them. I just don't think
we need to do that at this stage.