Output controller and model logs to archive
The following will tail the log files for the controller and any
model that uses the add_model function. This should make it easier
to see if there is a problem with the controller/model whilst
a test is running.

This does it's best to ensure we clean the pids up, but it's not
100% acurate. Also if the pid has been replaced with the same id, then
you're SOL'd.

For most if not all CI jobs this will be fine.