Create custom AMI image for deploy tests which test image id
Do this by moving the code to generate amis into includes. Also cleaning
it up, as there were a number of unnecessary steps (ssh keys, sec groups)

Use this library to create a custom ami on the fly, and then load this
constraint into our bundles

NOTE: This lib needs to support amd64 and arm64

Also, as a flyby, allow base and image-id combo. image-id insisted a
series was provided in bundles. However, now that bases have been added
these should be supported as well