export bundle default series properly
'default-series' no longer exists in configs. We parse it from
default-base. This meant that we were ignoring the default base/series
when exporting bundles. Fix this bug

Really we should be using a config method to query config items.
Parse the config, and load it's default base.

This the needs to be converted to a series to maintain compatibility

This meant small changes to tests, since our fake configs now need to
parse correctly.

There was one test (TestMixedSeriesNoDefaultSeries) which appeared to be
incorrect. Fix this