Fix potential leak
If we error out in Invoking the catacomb, we will leak a namespace
watcher.

Fix this by moving the creation of the namespace watcher inside our
watchers setup, and adding to it's catacomb immidately. Then re-assign
the watchers in channel

This ensures that the namespace watcher is for sure tied to our
watcher's catacomb, and that by the time the loop starts the watcher's
input channel is set