Prevent raising error prematurely in ModifyModelAccess
Fix LP 2028939.

Fixes a bug introduced in
https://github.com/juju/juju/commit/f30e2a0886df0d87e625d7c37fa27308920b30a2,
which changes the signature of HasPermission() from returning a (bool,
error) to returning just the error (that indicates the entity doesn't
have the permission for the given operation). It did, however, kept
the old error check to raise if there's an error coming from the
HasPermission itself, which is no longer valid, and causes the
ModifyModelAccess to prematurely raises an error if the user doesn't
have AdminAccess on the model, even if it has a SuperuserAcess on the
controller.