Raise error if something else (than permissions) has failed
This solves the issue in LP 2028939 (described in
e081275446ae248d040bae45e9c1d11552cd5bc2) in a way that doesn't throw
away but raises the error if it's not related to the permissions. If
it is an `ErrorEntityMissingPermission`, then we defer the decision to
raise until we check the `SuperuserAccess`.