Fix show cloud yaml support
The yaml support was broken in subtle ways. To perserve the old display
we add a new formatter that uses it. The yaml and json formatters work
correctly by outputting the right format to the stdout.

In addition we no longer write to standard out for yaml output. This
was breaking yaml in some decoders.

I think this is a much better approach rather than the janky yaml display
formatter.