Ensure mongo snap is running
The following ensures that the running of the mongo snap is done.
This failure can happen if we try to start a snap that's already
doing some pending work. As we're using the CLI directly we don't
have a way to check the error message nicely (unlike an API). This
means we have to just keep retrying until it works, or alternatively
we fail and bounce and try again.