Ensure the mongo service (snap) is installed
The following merges both the install and started functions. The
benefit of doing this is so that when we install it, we ensure that
mongo is actually started and running. Previously, we delayed that
when running in HA, and only forced it to start when attempting to
upgrade the database via the OpenState function. This caused problems
when entering HA. A race occurs when it expected that the mongo
service was running and how long it actually took to start. This
caused replicaset to fail to cluster because the underlying mongo
service wasn't actually running.

The solution is to just wait when installing it.
.