Remove external controller construction
The following removes the external controller construction from
the facades. We can now use the service factory for all the goodness.
The change makes the facades a lot more approachable and readable.

Ideally it would be nice to get rid of the ControllerDB, but we can't
do that just yet. Once state has been removed then it should be removed.