Work to remove controller config from APIHosts
The following starts to move the controller config from inside of
the state methods:

 - APIHostPortsForClients
 - APIHostPortsForAgents
 - SetAPIHostPorts

The consequences of this change means that the facades now have to
pass in the controller config at request time. The downside to this
is that we now have to ensure that common apiserver composible types
are now correctly wrapped. This is to ensure that we're not breaking
API compatibility for no reason.

There is a lot going on here, and it's not quite correct in terms
of tools type, which will be corrected in a subsquent commit.