Allow series and base in bundles if they match
Add a verification step into the client to ensure that series and base
match.

As such, we can treat series and bases as equivalent when they show up
precedence-wise.

Immidately combine into a base in core. As a helpful side effect, this
means we can minimise our dealing with series in this part of the code,
which will make it very easy to remove completely

Drop changes to core/bundle/changes/handlers.go in cherry-pick since we
are sticking with series in transit in 3.1