Use service worker factory for the controller config
The following passes the controller config from the service worker
factory via the dependency engine.

This allows us to directly inject the controller config nicely into
any worker without creating a service. This will be the standard practice
for how a controller config or for that matter any service will be
used in a worker.