Inject ctrl config model manager worker
This threads the controler config into the model manager, ensuring
that we correctly pull out the controller config for every model
restart. This could be why we're not seeing the model logs being
rotated correctly when creating a model, or when a model bounces
in the worker runner.

Like the peergrouper worker, this pushes the controller config from
the service worker factory via the dependency engine into the worker.
This pattern is starting to feel really nice.