Inject controller config to http server args
The following passes the controller config through the http server
args all the way to the authenticator. The authenticator is the
function that uses the controller config.

The controller config method now requires a context.Context to correctly
manage the lifecycle of the db request. We do this in two ways:

 1. We tie the controller config request to the tomb.
 2. We allow the passing in of the context to the request for the config.

For 1, we now make the http server args a real worker. This means that
we have a way to correctly tie each controller config request to the
underlying tomb. If the tomb dies we now correctly kill the request
at the earliest possible moment.

This is now complementary to 2, which means that if the ws connection
dies, we also close the request there. This marries both lifecycles
that the authenticator interacts with and should ensure a clean up
when required.