Remove SeriesSelector
Replace it with BaseSelector

This is part of the ongoing task to replace series with base

In most places, BaseSelector and SeriesSelector behave the same. There
are a few notable exceptions. BaseSelector prefers the latest LTS over
the charm's default, which results in some unit test failures.

Also, as a flyby, change NewCharmAtPath to use a base instead of a
series