Wire up tracing in API Server
The following pushes the tracing client all the way through the
many layers of the apiserver. It's very rough about how it all works.

The types found in the worker need to be pushed into core and the
rpc.Span is a bit weird, which we'll need to resolve later on.

The core of the solution is that the rpc connection gives everyone
a new context with a tracer and a span. Facades get it for free and
can then we can start just adding spans where required.

Currently there is no tagging and auto name creation, which would
make this more useful.

The controller config also needs resolving so you can turn this on
and off correctly. Along with removing the hardcode ip jaeger IP.