Add metrics to the changestream and dbaccessor packages
Adds new metrics for the changestream (both stream and eventmultiplexer
included) package, for monitoring current number of subscriptions,
errors, number of dispatched changes and watermarks.
Also, on the dbaccessor package a new metric was added for counting the
number of transaction retries, keeping a separate metric for total
number of transactions.