Move legacy tests to their own suite
I've broken out some tests in to their own legacy suite. This is
because they require a sync point that we're not willing to add
to the new dqlite infrastructure. Also, there is __no__ way to know
why your test failed, as these tests construct the whole environment
and assert on the outcome of a small output. These are tautological
by definition.

Instead the machine code should be broken into smaller composable
testable units. Each can then be verified independently. Then a
happy path integration test can be used to verify that the composable
parts correctly align.