golangci-lint doesn't like errors.IsNotFound
Presumably we only lint check files that are changing, and somehow that is
causing us to trip over these files for code that I did not change. Regardless,
I touched the file, I can clean it up a bit.

Switch from `errors.IsNotFound(err)` to `errors.Is(err, errors.NotFound)`