Pass in controller config for migrations
We require the controller config for migrations, in this change
we now pass it explicitly from the domain rather than the state.

Doing this is only possible because the controller config is a global
collection. There is still another controller config call site within
state that also needs to be purged, but we can do that in another
change request.