Support bases in bundles
This commit backports _some_ of the changes from the PR
https://github.com/juju/juju/pull/15725

However, instead of converting to bases at the edge, we convert to
series at the edge. This is because there are a lot of changes in later
versions that using bases throughout depends on which are not present in
3.1, which would require expensive backports

We also do NOT backport changes to GetChanges and GetChangesMapArgs, as
they are not required to backport this feature. This simplifies the
change significantly