Context-aware common API methods
The following introduces context.Context to all public API methods.
We're doing this for 2 reasons:

 1. We require that all requests to the domain level service be
    cancellable (or aware of being cancelled). This is to ensure
    we don't block the dbaccessor from restarting or from being
    blocked at any point.
 2. As an side we can start to pass tracing through all the layers
    of the stack. The standard practice is to do this via the
    context.Context package. This makes this a lot simpler.

The work here is repetitive and long winded. Any public method now
gets a context.Context as a first argument. Tests are fixed to pass
a context.Background(). There are a few cases of context.TODO(), which
will require another pass to remove those. The work will involve tying
the context to the worker tomb if there is one, or passing in some
sort of managed wrapper for that context.

The fallout of this work is that ModelConfig and CloudSpec also become
context aware. This is unfortunate, as we had to modify the state
package, yet this is a side effect of how we use the client, state
and other various implementations interchangably.