Tracing worker
The following is the tracing worker. This worker will manage the
lifecycle of a tracked tracer. A traced tracer is a wrapper around
a client for a given namespace (model or controller namespace), so
that if a tracer client fails, it doesn't bring down every other
tracer. As we're using the runner package, a new client will be
spawned and a tracer can be selected again.

This brings in a new dependency: go.opentelemetry.io/otel. This is
the default standard for open telemetry, although it does have a
few rough edges.

 1. There is no logging support (assoicating logs with spans)
 2. There is a global instance logger, which we will have to work around

Hopefully point 1 is actively being worked on and might arrive in
the not to distant future.

As this is just setting up the worker, there isn't any manual tests
to add to this PR, that will come later. For now it should bootstrap
and react to the changes in the agent/controller config.

The implementation mirrors the dbaccessor, including the new error
catacomb.ErrDying support.