Rework domain/controllerconfig package
The controller config package in domain was wrong. We removed required
fields from the controller config and we coerced on the way in and on the
way out, each with different outcomes.

The following fixes that by not validating on the way in, instead we just
serialize (encode) all the config values to strings. Then on the way out
we just coerce them when required. This ensures that the controller config
map always has the right values at the right time.