Replace SeriesForCharm with BaseForCharm everywhere
This means we can remove SeriesForCharm, another occurence of series
dropped.

This function was only used to validate bases, so dropping it was as
simple as using BaseForCharm and dropping a conversion layer

In these validation functions, I have also removed a fallback of the
supported series to the series in the charm url. I believe this is valid
because this is old functionality hanging over from a previous version of Juju