Watcher test harness
The following implements a watcher test harness to allow integration
tests with the real change stream and event multiplexer. To prevent
circular imports issues we need to move the code outside of the worker
package. Doing this, then makes it possible to create a watchable db
controller/model suite.

I've reworked the credential watcher and tests associated so it does
actually use this new test harness. In order to facilitate this change
we also need to add assert changes, when allows any dispatches to be
witnessed for a given time frame, thus giving a clean state for future
watcher changes.