Retry transactions when entering scope of a relation unit
When using server-side transactions, mgo retries (hard-coded 3 times)
transactions that fail with WriteConflict error, and after that it
aborts the (mgo) transaction.
These retries are seen many times on multiple scenarios in juju, and therefore
the business logic decides on each case how to treat the (mgo
server-side) aborted transactions.

In the case of EnterScope, it is necessary to retry the (mgo server-side)
transaction (which includs retrying the mongodb transaction) because EnterScope
is called concurrently (via the EnterScope facade) and therefore WriteConflict
errors are transient errors. By calling Run(jujutxn.TransactionSource)
instead of RunTransaction(ops []txn.Op) we make use of the retries
applied to the function (jujutxn.TransactionSource) passed to Run(...).

Note: This last observation means that the true solution would be
sequential transactions for EnterScope, but since each relationunit has
it's own runner, this would mean a huge change.

This patch fixes a bug (https://bugs.launchpad.net/juju/+bug/2031631) on
which the relation unit doesn't enter scope in the case multiple units
and applications are present before creating a relation on them.