Remove redundant expensive build jobs
These tests are meant to test that we don't accidently break the
building of juju(d) because of some architecture dependant issue.
Initially designed as a smoke test and therefore meant to be quick.
In addition to that, they're supposed to mirror production builds via
the snap infrastructure. Unfortunately, powerpc64 and s390x are
neither quick to build nor replicate production builds.

Just to be clear arm64 and amd64 don't replicate production building
via snaps, but at least they are quick and do at least service as
smoke tests.

We do have additional jenkins jobs to verify that do the do correctly
build on all architectures in the correct way, so this should be
fine and release some github runners to be able service other jobs.