Ensure we subscribe inside the loop
The following updates the eventmultiplexer to ensure that the
subscription is added in the loop. This ensures that the race during
creating a subscription and then return the sub, but it's not actually
registered yet. Thus you can act upon a sub which is not yet available.

The fix is to push everything into the loop and only return a subscription
once everything has offically been subscribed.