Add additional stream tests
The follow adds additional test to the stream based on a discussion
with how the change stream works.

In testing the code I noticed that the metric for inserting a watermark
is in the wrong place. We should only record the metric on success,
otherwise it will seem like we're recording on every ask. This isn't the
case as there might be no additional work to be done.