API Client only use windowed facades
The following forces the api client to use specific versions of the
api facades. Doing this by offering a slice of versions that the
client can use based on the apiserver api server facades.

This fixes the problem for migrations where prechecking is done on
both client and server because we don't know which actual client
facade versions we support. Once this lands, we can drop the checks
on the client and server for major versions and just instead check
for which the facade version supports.