Remove of V1 ApplyLease Facade
The following removes the v1 ApplyLease Facade as the v2 version
just passes a JSON payload that we be decoded once. The v1 version
will double decode, once in json and the second in yaml for the
command.

We no longer supper v1, so it should be safe to remove.