Prevent bouncing worker for action retrieval
Prevent the bouncing of the worker if we can't find the action. This
causes a lot of churn in the dependency engine, even if we can't get
the action. Instead, just log the error and move along.

We have taken the decision that workers should only be restarted on
critical or non-retryable errors. This clearly falls into this
category.