Return ErrTracerDying from GetTracer
The following ensures that we don't return catacomb.ErrDying. That
error doesn't belong outside of the loop, so we should ensure not
to pass it to anyone else. Doing so can be dangerous, as it will
cause other catacombs to fail with an error (boucing the catacomb)
instead of closing gracefully.