Fix storage provisioner tests
The following puts the right tests in the right files to make it
a lot easier to work with. In addition, this puts the db front and
center in these tests and removes the hack of setting a fake one and
then only using the real one if the fake one doesn't exist. Now
we're consistent.