Adds new method WithPreferredCloudLocalAddressOption to NodeManager.
This will be used during bootstrap if the behaviour is specifically
requested, to use a unique local cloud address if one is available.

This will prevent churn during a future enable-ha invocation because
Dqlite will not have to be rebound.

If a unique address is not found, we fall back to using the loopback
device.