Attach a tracer to a context
The following allows the adding and retrieval of a tracer to a context.
This is useful for when we don't want to explicitly pass a tracer
through the many layers of the stack. Normally, I would advocate to
be explicit, but as this is self contained piece, then we don't
necessarily need to be concerned about this. If the tracer does leak
outside of this package and into code, then it's at that point we
need to re-evaluate that descision.