Start a trace at the rpc server layer
By ensuring that we start a trace at the rpc layer, we can then
workout if we're getting one from a worker or from a client. At the
moment we don't take into account a trace from a worker, that'll come
later. For now just spawn a new one and pass it through to every call.