Push the concept of service name forward
The following pushes the concept of the service name forward throughout.
This means that you no longer care about the model uuid directly. To
search for a given model, you'll need to search via a tag. Tracing is
concerned about inter service relationships. Changing this makes this
aware of that.