Remove client version checking
The following removes client version checking and restriction. The
problem with this system is that there are two sources of truth and
there is a competing idea; API facade versioning.

The two sources of truth are where you're coming from and where you're
at. Prior to this change we just restricted clients based on a
client version and we did client side checks of versioning. Except the
old client then needed to know about future versions. I removed that
check previously. The reason we want to remove this, is that in order
to discount a client that is perceived "bad" both the old and new
client needed to be aware of it. Requiring a sync'd release.

The alternative is the API facade versioning. Which states that if there
is a version available just use that. With the recent change to have
a range of versions rather than just an max int available, we can ensure
that the client won't allow for older versions. The api client can
now ask for an intersection of versions valid for it.

Doing this means, if you want to block off an older client, you only
need to change the current implmentation by bumping all the facades
you want to restrict and removing the older ones. Thus providing only
one location for client side restrictions.

This works across minor, major and patch releases without the embedding
of a version anywhere. Genius! Just as the system was intended.