Fix for LP1988556, broken refresh by revision for charmhub.
This is a patch based on a work around of using the switch flag with the
new revision in the charm url for charmhub. The work around is
suboptimal, as we want to not encourage charm url use by users directly.
It's going away.

A real fix will be more possible in juju 3.1 which does not contain
charmstore charms. Juju needs to update its use of the charmhub refresh
action for the refresh endpoint. However to share deploy/refresh code
with charmstore easily, it's necessary to use ResolveCharm too.
ResolveCharm does not differentiate between deploy and refresh
operations as necessary for charmhub charms, but not charm store charms.