Add controllercharm facade
This is the first part of ongoing work to integrate the Juju `controller` charm with `prometheus-k8s`.

The controller charm needs to create/remove user credentials on the fly, to pass to Prometheus to access Juju's metrics endpoint. Here, we define a new `controllercharm` facade, which contains Add/RemoveMetricsUser methods to be used for exactly this purpose.

Future work will define a method for the controllercharm to access this facade.