Ensure API Client can call 3.x API Facades
The following ensures that with the new API clients 4.x can call
3.x API facades. This is required now that we're dropping support
for restricting clients based on version numbers.

Validation of client facade versions can then ensure that we correctly
don't drop support for vital paths. Luckily, the dropping of ModelManager
v9 was dropping away series support, but newer clients could talk to
v9 facades via bases. If this was different, we would need to ensure
that 4.x clients could still talk to 3.x latest facades.

The need to remove old facades is actually quite difficult, as each
facade is a collection of methods and not a singular method. So viewing
them as whole is fundamental. Thus attempting to drop any facade
without a major release is prohibitied and even then we should validate
if it's even possible.

The code changes are simple, allow the client to pick a 3.x client
version and add a test that locks in the change. Changing these tests
will ensure that it can't just be done willy nilly.