Remove precheck validation for migrations
The following removes the precheck validation for migrations. We
should never be validating based on a client version for migrations.
API facades are the true way to do this.

The rabbit whole that you inevitably end up with, is that older
jujus will at some point need to know about the next major version
because we allow .0 releases to jump to major versions.