Improve error message for missing facades/methods
The following attempts to improve the error messages for missing
facades and methods. It wraps the NotImplemented error type from
the APICall to give a more bespoke error message.

Unfortunately, depending on the facade you might still get a
old error message from an older controller. In those cases we provide
some context for that with a resolution path forward.