Clear out the snap channel for legacy tests
The legacy tests are clearly testing too much. There is no way
you know when it blows up what broke something without spending
hours/days on it. This is the third time in 3 weeks that I've tried
to bring these things back to life.

They are valuable, just not as unit tests. I think we should remove
them and migrate them to integration tests.