Remove the last occurences of ComputedSeries
Remove this function from core

This also meant dropping the fallback to aquring a charm's supported
OSes from the charm url. This is not a supported feature anymore which
it turns out a larghe number of our test charms were depending on.

Add manifest.yamls to a number of our test charms