Moving ServiceFactory interfaces to internal.
With this PR we are moving the ServiceFactory interfaces defined in the
worker outside to an internal package. By doing this we can avoid the
layer violation of having a bunch of Juju components referencing types
out of the worker and subsequently getting the package imports from the
worker.

This is needed for future PR's on dqlite that are introducing new
testing suites for the ServiceFactory.

This change is self contained and only deals with the interfaces defined
in the worker.