Inject controller config from agent bootstrap
The following injects the controller config into the agentbootstrap
removing the reliance on state package. Unfortunately, we have to
return a txn runner from bootstrap dqlite. This ensures that the right
values are correctly returned that is consistent with what is within
the db. This does mean that bootstrap isn't as pure as we'd like, but
it does solve the problem of ensuring what is in the database is
the source of truth.

There will become more of a requirement for this to happen in the
future with later requests to the underlying db. We might have to
actually create a full service factory, but it becomes harder to
do as we don't have a change stream which the service factory
requires. We might have our modelling off slightly here.