Remove gloabl init state function
The following removes the init state function that was patched in
the tests, in favour of setting one directly on the command. We can
then offer a way to push that in as a dependency.

Doing this makes it more obvious about what the bootstrap command
requires. This cleans up the need to patch the value at runtime.